Make uv_write_t's result field an `int`, like before
This field didn't actually need to be an `ssize_t`, so we can make it an `int`
to avoid changing the struct size.

+++ include/uv/unix.h
@@ -259,7 +259,7 @@ typedef struct {
   unsigned int write_index;                                                   \
   uv_buf_t* bufs;                                                             \
   unsigned int nbufs;                                                         \
-  ssize_t result;                                                             \
+  int result;                                                                 \
   uv_buf_t bufsml[4];                                                         \
 
 #define UV_CONNECT_PRIVATE_FIELDS                                             \
